iemgui_dialog: properly escape all strings
use a helper-function to create an iemgui_dialog
(the sprintf() is *really* hard to read, so now we have a simple
function with understandable parameters (names!).
a decent code editor will give us online help about the parameters,
making it easy to get the call right.

also: formatting fixes need to be done only once...)

Closes: https://github.com/pure-data/pure-data/issues/1562
Closes: https://github.com/pure-data/pure-data/issues/622